Look up geo IP headers in rawHeaders and log them
I checked the serverless function logs and it seems as though the structure of the rawHeaders array is an alternation of keys and values. I look up the indices of the keys of interest and pull the corresponding values shifted by one. I also log request.headers in case the typings do not show the members for these headers but they are available at runtime. If I rule this out, I will stick with the above, but if I find members for them, I will rewrite the code to treat request.headers as any and pull the respective members from that.