Stop logging country, region, city to investigate bad request
All of the sudden, Supabase started returning 400s for the API hit log insertion calls, but not middleware or UI calls. I think it may have to do with the types of the country, region and city values, but in the serverless function logs they appear as strings... Maybe that's just an implicit toString call. I think the solution will be to call toString on the expression generating the values, but before I do that, I want to check if removing them makes the Supabase client work again.